add type constraint

+++ app/Console/Kernel.php
@@ -36,11 +36,11 @@ class Kernel extends ConsoleKernel
     {
         $schedule->command('invitation:cleanup')->daily();
         $schedule->command('remind:score')->daily()->skip(function () {
-          return ((int) date('i') > env('CRON_BREAKPOINT', 40));
+          return ((int) date('i') > (int) env('CRON_BREAKPOINT', 40));
         });
 
         $schedule->command('remind:game')->daily()->skip(function () {
-          return ((int) date('i') > env('CRON_BREAKPOINT', 40));
+          return ((int) date('i') > (int) env('CRON_BREAKPOINT', 40));
         });
         $schedule->command('queue:restart')->hourly();
         $schedule->command('queue:work --tries=3 --sleep=10')->hourly()->withoutOverlapping();
